updated quantity, minQuantity to be doubles instead of int. Api call and other code reflects this

+++ lib/models/inventory.dart
@@ -1,7 +1,7 @@
 class Inventory {
   final String name;
-  final int quantity;
-  final int purchaseQuantity;
+  final double quantity;
+  final double purchaseQuantity;
   final int costPerPurchaseUnit;
   final String unit;
   final String notes;
